fix
git-svn-id: https://everest.teltel.com:8443/repos/misc/9x9dev/trunk/nncloudtv@4724 a0a5f996-2bb4-401c-8778-395c3c01700f

+++ src/main/java/com/nncloudtv/web/api/ApiPoi.java
@@ -927,7 +927,7 @@ public class ApiPoi extends ApiGeneric {
         if (verifiedUserId == null) {
             unauthorized(resp);
             return null;
-        } else if (verifiedUserId != ownerUserId) {
+        } else if (verifiedUserId.longValue() != ownerUserId.longValue()) {
             forbidden(resp);
             return null;
         }
@@ -1038,7 +1038,7 @@ public class ApiPoi extends ApiGeneric {
         if (verifiedUserId == null) {
             unauthorized(resp);
             return null;
-        } else if (verifiedUserId != ownerUserId) {
+        } else if (verifiedUserId.longValue() != ownerUserId.longValue()) {
             forbidden(resp);
             return null;
         }
